Rewrite category updating implementation in the DB
Previously, a new category node would be formed, and all relationships (symbols,
articles, etc) would be recreated onto this head node. This didn't work when
many symbols connected to a category (causing a stackoverflow in neo4j). A more
performant way of keeping the head node as-is, and simply creating a revision
node containing the node's content before the update, is now done.